requirements.txt
asgiref==3.7.2
Django==4.2.7
django-browser-reload==1.12.1
django-tailwind==3.6.0
Pillow==10.1.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.0
django-environ==0.11.2
django-restframework==0.0.1
djangorestframework==3.14.0
psycopg2==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
